fixed the height
This is the height I used with the post function I wrote and it works pretty well.  I thought you'd be working on making the game end when one player scores 7 points since the post is kind of more related to what I'm supposed to be doing. That's is kind of why I was hoping we'd be able to get together to work on this, it seems like we've been repeating a lot of code.